# experimental/gateway/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "additionalProperties": true,
        "description": "Every error answer (4xx and 5xx).",
        "properties": {
          "error": {
            "description": "What went wrong, in words a person can act on",
            "examples": [
              "no upload u7"
            ],
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LogPage": {
        "additionalProperties": true,
        "properties": {
          "loggers": {
            "description": "Every logger name seen, sorted (for a filter menu)",
            "items": {
              "type": "string"
            },
            "title": "Loggers",
            "type": "array"
          },
          "offset": {
            "description": "Byte offset to pass back as `after` for only newer records",
            "title": "Offset",
            "type": "integer"
          },
          "records": {
            "description": "Matching records, oldest first",
            "items": {
              "$ref": "#/components/schemas/LogRecord"
            },
            "title": "Records",
            "type": "array"
          },
          "runId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The run read (`latest` resolved)",
            "examples": [
              "20260101-120000-unitree-go2"
            ],
            "title": "Runid"
          }
        },
        "required": [
          "runId",
          "records",
          "offset",
          "loggers"
        ],
        "title": "LogPage",
        "type": "object"
      },
...
      "SkillCallRequest": {
        "additionalProperties": true,
        "properties": {
          "args": {
            "additionalProperties": true,
            "description": "Its arguments, by its `params` schema",
            "examples": [
              {
                "command_name": "FrontJump"
              }
            ],
            "title": "Args",
            "type": "object"
          },
          "module": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Only this module's skill of that name",
            "title": "Module"
          },
          "runId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Refuse unless this is the running run (default: whichever runs)",
            "title": "Runid"
          },
          "skill": {
            "description": "The skill's name, as GET /dimos/skills lists it",
            "examples": [
              "execute_sport_command"
            ],
            "title": "Skill",
            "type": "string"
          }
        },
        "required": [
          "skill"
        ],
        "title": "SkillCallRequest",
        "type": "object"
      },
      "SkillCallResult": {
        "additionalProperties": true,
        "properties": {
          "blueprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "That run's blueprint",
            "examples": [
              "unitree-go2"
            ],
            "title": "Blueprint"
          },
          "content": {
            "description": "Its whole answer as MCP content blocks (`{type: text, text}`, `{type: image, data, mimeType}`)",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Content",
            "type": "array"
          },
          "module": {
            "description": "Its module",
            "examples": [
              "UnitreeSkillContainer"
            ],
            "title": "Module",
            "type": "string"
          },
          "ok": {
            "description": "False when the skill failed (`text` says why)",
            "title": "Ok",
            "type": "boolean"
          },
          "runId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The run it ran in",
            "examples": [
              "20260101-120000-unitree-go2"
            ],
            "title": "Runid"
          },
          "skill": {
            "description": "The skill called",
            "examples": [
              "execute_sport_command"
            ],
            "title": "Skill",
            "type": "string"
          },
          "text": {
            "description": "Its answer's text",
            "examples": [
              "'FrontJump' command executed successfully."
            ],
            "title": "Text",
            "type": "string"
          },
          "via": {
            "description": "`rpc` (dimos's module RPC) or `mcp` (the run's McpServer, for a skill that holds a capability)",
            "examples": [
              "rpc"
            ],
            "title": "Via",
            "type": "string"
          }
        },
        "required": [
          "skill",
          "module",
          "runId",
          "blueprint",
          "via",
          "ok",
          "text",
          "content"
        ],
        "title": "SkillCallResult",
        "type": "object"
      },
...
      "Upload": {
        "additionalProperties": true,
        "description": "One upload in the queue.",
        "properties": {
          "bytesDone": {
            "description": "Bytes done in this phase",
            "title": "Bytesdone",
            "type": "integer"
          },
          "bytesTotal": {
            "description": "Bytes in this phase (0: indeterminate)",
            "title": "Bytestotal",
            "type": "integer"
          },
          "createdAt": {
            "description": "When it was queued (Unix ms)",
            "title": "Createdat",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why it failed, or that it waits for a login",
            "title": "Error"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "not_logged_in, network, quota, file_missing or failed",
            "examples": [
              "not_logged_in"
            ],
            "title": "Errorcode"
          },
          "etaSeconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Time left in this phase (null for the first second)",
            "title": "Etaseconds"
          },
          "finishedAt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "When it finished (Unix ms)",
            "title": "Finishedat"
          },
          "id": {
            "description": "The upload's id",
            "examples": [
              "u3"
            ],
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "recording, video, pointcloud, log or blob (null: from the file)",
            "title": "Kind"
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Its page in the Dimensional console, once done",
            "title": "Link"
          },
          "log": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The log with the details of a failure",
            "title": "Log"
          },
          "name": {
            "description": "Its file name",
            "examples": [
              "walk.mcap"
            ],
            "title": "Name",
            "type": "string"
          },
          "notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Something to tell the user, e.g. about the quota",
            "title": "Notice"
          },
          "path": {
            "description": "The recording",
            "examples": [
              "/home/me/.dimos/recordings/walk.mcap"
            ],
            "title": "Path",
            "type": "string"
          },
          "phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "While uploading: preparing, compress, hash, upload, then finishing",
            "examples": [
              "upload"
            ],
            "title": "Phase"
          },
          "rateBps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Smoothed speed in bytes a second",
            "title": "Ratebps"
          },
          "robotId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The robot id it's tagged with",
            "examples": [
              "go2-lab"
            ],
            "title": "Robotid"
          },
          "size": {
            "description": "Its size in bytes when queued",
            "title": "Size",
            "type": "integer"
          },
          "skipped": {
            "description": "The cloud had it already",
            "title": "Skipped",
            "type": "boolean"
          },
          "startedAt": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "When its upload started (Unix ms)",
            "title": "Startedat"
          },
          "state": {
            "description": "queued (first in, first out), uploading (one at a time), then done, failed or cancelled",
            "enum": [
              "queued",
              "uploading",
              "done",
              "failed",
              "cancelled"
            ],
            "title": "State",
            "type": "string"
          },
          "uploadId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The cloud's id for it, once done",
            "title": "Uploadid"
          }
        },
        "required": [
          "id",
          "path",
          "name",
          "size",
          "robotId",
          "kind",
          "state",
          "phase",
          "bytesDone",
          "bytesTotal",
          "rateBps",
          "etaSeconds",
          "uploadId",
          "skipped",
          "notice",
          "error",
          "errorCode",
          "log",
          "createdAt",
          "startedAt",
          "finishedAt",
          "link"
        ],
        "title": "Upload",
        "type": "object"
      },
...
      "UploadList": {
        "additionalProperties": true,
        "properties": {
          "uploads": {
            "description": "The queue, in order",
            "items": {
              "$ref": "#/components/schemas/Upload"
            },
            "title": "Uploads",
            "type": "array"
          },
          "waitingForLogin": {
            "description": "An upload found no cloud login: the queue waits for one",
            "title": "Waitingforlogin",
            "type": "boolean"
          }
        },
        "required": [
          "uploads",
          "waitingForLogin"
        ],
        "title": "UploadList",
        "type": "object"
      },
...
      "UploadRequest": {
        "additionalProperties": true,
        "properties": {
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "recording (default for .mcap/.db), video, pointcloud, log, blob",
            "examples": [
              "recording"
            ],
            "title": "Kind"
          },
          "path": {
            "description": "the recording's absolute path (a /recordings entry's path): an .mcap or a .db",
            "examples": [
              "/home/me/.dimos/recordings/walk.mcap"
            ],
            "title": "Path",
            "type": "string"
          },
          "robotId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "robot id to tag it with",
            "examples": [
              "go2-lab"
            ],
            "title": "Robotid"
          }
        },
        "required": [
          "path"
        ],
        "title": "UploadRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The dimos gateway's HTTP API: blueprints, global config, runs and their logs, events, Dimensional cloud uploads,\ndiscovery (blueprints, modules, message types), docs, extras and jobs.\n\ndimOS Desktop starts the gateway (`python -m experimental.gateway`, on a unix socket) and forwards `/dimos/...` to it unchanged.\nEvery answer is JSON unless the operation says otherwise; every error is `{\"error\": \"<message>\"}` (ErrorResponse).\nState is HTTP; changes are events, published on zenoh at `<ns>/dimos/events/<type>` (the `events` tag).\n\nExtensions: `x-family: dimos` on every operation; `x-agent: true` marks what Desktop's agent finds through\nsearch_endpoints; `x-mcp-tool` names an MCP tool that does the same; `x-zenoh-key` on an event schema is its key.\n",
    "title": "dimos gateway",
    "version": "1.18.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/dimos/runs/{runId}/log": {
      "get": {
        "description": "Reads `<logs>/<runId>/main.jsonl` off disk (the checkout's logs/, then the library install's; the last 4 MB at most). Without `after`: the last `limit` matching records; with `after`: every matching record past that byte offset (tailing). An unknown run answers no records. No side effects.",
        "operationId": "get_dimos_runs_runId_log",
        "parameters": [
          {
            "description": "run id or latest",
            "in": "path",
            "name": "runId",
            "required": true,
            "schema": {
              "description": "run id or latest",
              "examples": [
                "latest",
                "20260101-120000-unitree-go2"
              ],
              "title": "Runid",
              "type": "string"
            }
          },
          {
            "description": "byte offset from an earlier answer's `offset`: only newer records",
            "in": "query",
            "name": "after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "byte offset from an earlier answer's `offset`: only newer records",
              "title": "After"
            }
          },
          {
            "description": "minimum level: debug, info, warning, error",
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "minimum level: debug, info, warning, error",
              "examples": [
                "warning"
              ],
              "title": "Level"
            }
          },
          {
            "description": "text to match",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "text to match",
              "title": "Q"
            }
          },
          {
            "description": "at most this many records (default 1000; ignored with `after`)",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "at most this many records (default 1000; ignored with `after`)",
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogPage"
                }
              }
            },
            "description": "`{ runId, records: [{ timestamp, level, logger, event, extra, raw }], offset, loggers }`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the gateway refuses (the message says which)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The gateway couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "A run's structured log (main.jsonl): records with level, logger, event",
        "tags": [
          "logs"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
    "/dimos/skills/call": {
      "post": {
        "description": "Calls `<module>/<skill>` over dimos's module RPC, as the coordinator calls a module's start. A skill that holds a capability (`uses`) goes through the run's McpServer instead when one answers (`via: mcp`), so its agent's capability locks cover it: it waits or is refused while another skill holds one. A background skill answers at once. This acts on the robot. Waits up to 300 s. `ok` false when the skill itself failed. 400 for a missing or unknown argument (checked against `params`) or a name two modules share without `module`, 404 when the running blueprint has no such skill (or `runId` isn't it), 409 when nothing runs, 500 when the MCP call failed or it didn't answer in time.",
        "operationId": "post_dimos_skills_call",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SkillCallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SkillCallResult"
                }
              }
            },
            "description": "`{ skill, module, runId, blueprint, via, ok, text, content }`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the gateway refuses (the message says which)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such thing (the message names it)"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict with the current state (the message says what it is)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The gateway couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Call a skill of the running blueprint and wait for its answer",
        "tags": [
          "skills"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
    "/dimos/uploads": {
      "delete": {
        "description": "Removes every done, failed or cancelled upload from the list (what is in the cloud is still remembered, see /dimos/uploads/uploaded) and answers the queue.",
        "operationId": "delete_dimos_uploads",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadList"
                }
              }
            },
            "description": "`{ uploads: [Upload], waitingForLogin }`, without the finished ones"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The gateway couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Clear the finished uploads (done, failed, cancelled) from the list",
        "tags": [
          "uploads"
        ],
        "x-agent": false,
        "x-family": "dimos"
      },
      "get": {
        "description": "The queue in order, and whether it waits for a cloud login. No side effects.",
        "operationId": "get_dimos_uploads",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadList"
                }
              }
            },
            "description": "`{ uploads: [Upload], waitingForLogin }`"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The gateway couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "The Dimensional cloud upload queue: each upload's state, progress, speed, time left and error",
        "tags": [
          "uploads"
        ],
        "x-agent": true,
        "x-family": "dimos"
      },
      "post": {
        "description": "Adds the recording to the end of the queue (saved, so it survives a restart) and answers its upload, queued; one already queued or uploading for that path is answered instead. Needs a cloud login: without one it waits. 400 when the path isn't an absolute path to an existing dimos recording (an .mcap, or a .db dimos recorded).",
        "operationId": "post_dimos_uploads",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Upload"
                }
              }
            },
            "description": "`Upload`"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad request: a missing or malformed body or parameter, or a value the gateway refuses (the message says which)"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "The gateway couldn't do it: a child process, launch, stop or cloud call failed (the message says why)"
          }
        },
        "summary": "Upload a recording (.mcap or .db) to Dimensional cloud: it joins the queue (one at a time)",
        "tags": [
          "uploads"
        ],
        "x-agent": true,
        "x-family": "dimos"
      }
    },
...
  },
...
}